Add PostgreSQL service health check
When setting up the project for the first time, `update` can sometimes attempt
to apply database migrations before the PostgreSQL database has been
initialized. This change set adds a `pg_isready` health check other services can
depend on. In this case, the Django service is configured to depend on the
database being in a health state before it attempts to apply migrations.