Change user model to use email instead of username
In order to omit the username field entirely we have to subclass
`AbstractBaseUser` instead of `AbstractUser`, which also omits several
other helpful fields, and requires using a custom manager class.

I copied all of the fields (except username) from `AbstractBaseUser`
to try and keep us as close to the standard Django user as possible.

Fixes #47