Fix Angular development environment
Work toward an approach where all of the Node dependencies are
encapsulated by the container image. While building the image, use
changes to `package.json` as a way to indicate that the image needs
rebuilding.

For development, only volume mount the `src` directory inside of the
Angular project. This avoids volume mounting on top of files or
directories that were created as part of the image building process.
This change likely makes it more difficult to see changes made to files
outside of the `src` directory, but it looks like that may be an
acceptable trade-off. Changes made to files outside of `src` will
require a container image rebuild via `update`.