Break PlanItLocation out from org model
- Add helper on PlanItUser model to get location
- Fix tests with appropriate mock
- Only update location info when created
- Implement OrganizationSerializer create/update methods
- Remove uniqueness constraint on location.api_city_id

Removing the constraint was the lesser of two evils while
we're interating quickly on the definition of this object.
See:
- https://github.com/encode/django-rest-framework/issues/2996
- https://medium.com/django-rest-framework/dealing-with-unique-constraints-in-nested-serializers-dade33b831d9
Handling nested serializers with uniqueness constraints
requires overriding the default validators generated for the
entire serializer which is super inelegant. No harm to us for
now if what are essentially duplicate planitlocation objects
get created somehow.