Remove quotes from FLAKE8_ARGS
When executing `test` locally, the value of `FLAKE8_ARGS` is set to an
empty string. Because that variable is quoted when referenced, it causes
a trailing `''` in the command parameters:

  docker-compose ... --entrypoint flake8 django --tee ''

This residue is causing `flake8` to malfunction. Unquoting the variable
leads to no residue. This did not manifest itself in CI because the
value of `FLAKE8_ARGS` is set to something other than an empty string in
that environment.