Shift violations.xml output to src directory
The root of the working directory inside of the Node.js container is not
volume mounted, which prevents writes that occur from within the context
of the container instance from propagating out to the virtual
machine/NFS file system.

Since the `src` directory is volume mounted, writes within it do
propagate outward, allowing Jenkins to read `violations.xml`.