Fix ExpressionChanged error by delaying setting models list
Likely exposed by ClimateModel list being returned immediately
from the cache, causing the following steps to occur:
1. Dataset changes in ModelModalComponent
2. Triggers model list API request
3. ModelModalComponent.onModelsChanged fired
4. IndicatorChart.models is set from change event
5. Triggers onChanges for ModelModalComponent.models

Before, the model list API request happening asynchronously
was enough to break this chain of events.

We can fix the error by wrapping the update to IndicatorChart.models
in a setTimeout, which delays checking until next cycle.

This is a temporary fix, when:

https://github.com/azavea/climate-change-components/issues/28

is addressed, this fix should be removed as well.