Fix occasional error on autocomplete select
The order in which the item select and item blur callbacks fire is apparently indeterminate.
Do validation on blur with a slight delay, in case an item was selected from the dropdown,
but the form value has not updated before the blur event.

Fixes issue where sometimes clicking on an autocomplete suggestion would result in a
form error and blank field without closing the dropdown.