Updates climate-change-components and Angular CLI dependencies
We modified climate-change-components to use peer dependencies instead of
normal dependencies out of a desire to reduce the number of duplicated
dependencies in this project.

To that end I also updated the Angular CLI to version 1.4.10, which fixes
a recurring issue we've encountered with an incorrect peer dependency.

With these changes in place, I was able to flatten RxJS from 3 versions to
1 and core-js from 2 versions to 1 ð

Closes azavea/climate-change-components#30