Link to homepage for login for new, activated user
Update the login link on the user activation page to point to the app
homepage instead of the django-registration provided login page. That
login page is not integrated into the app.

Also, remove the login header link from the registration completion page
because this links to the django-registration provided login page.

Refs #210