Add link to download CSV export on dashboard
Add a service to allow for an asynchronous download of a file. This
approach is based on the implementation in Model My Watershed
(https://github.com/azavea/model-my-watershed-zappa/blob/develop/src/mmw/js/src/core/utils.js#L415).
A similar approach is also commonly recommended for use within Angular.

The location of the download link is temporary, and is for demonstration
purposes.

Refs #409