Fix validation check for typeahead fields
Using a setTimeout to delay the validation event on certain typeahead
fields allowed the user to proceed to the next step of the wizard with
an invalid field value. Removing the setTimeout call fixes the issue
described in #495.

The setTimeout was introduced to fix an issue (see the commit
message of 42c0576a30138503b67e0499b4a8f3285071c423). However, I believe
that issue is addressed by using the typeahead name-spaced version of the
blur event.

Refs #495