View risk link should go to last step of wizard
When the user clicks the view link for a risk on the assessment overview
page, it now goes to the last step of the risk wizard. This is opposed
to the edit link, which goes to the first step of the risk wizard.

To achieve this, a query parameter that indicates which step to go to is
added to the view link.

Refs #520