Remove object list endpoints for users and orgs
At present, any user could access the details of every user, which
could be a security risk. Therefore, the endpoints that list the details
of every user and organization are removed here. These endpoints are
currently not utilized in the app.  The tests are also adjusted to match
the new API behavior.

Refs #206