Redirect activated user to new login page
When a user activates their account, they are now redirected to a login
page within the Angular app. Previously, they were routed to a page
provided by the Django registration module.

This login page repurposes the login modal that is used on the marketing
page. If a logged in user tries to access the login page, they are
redirected to the dashboard. Also, the login link on the marketing page
now goes to the login page instead of opening the modal.

Refs #605