Close model modal when parent modal is closed
If the parent modal of the indicator charts is closed, close the model
modal. Because the parent modal isn't a direct parent of the indicator
chart, we check the visibility of the chart each time Angular checks for
changes to components. If the chart isn't visible anymore, close the
model modal. This ensures the model modal isn't left open when the chart
is closed, which 1) left an extra modal backdrop open over the app and
2) caused additional problems when the indicator charts were reopened.

Refs #660