Scroll to top when entering next wizard step
When proceeding to the next step in a wizard (used when setting up a
plan, selecting risks, and assigning actions), scroll to the top of the
page. This ensures users see the complete form on each step, and that
the scroll position of the previous step has no effect on the next step.

Refs #726