Show city and state in city selection menu
Helps disambiguate cities in different states with the same name. Also,
the admin (state) field was being searched, which lead to confusing
results because the state value was not displayed anywhere.

This is done by modifying the city results that come back from the API
and adding a new field that contains both the city and state name. Tip
of the hat to @maurizi for the solution.

Also, remove unused function.

Refs #762