Add error message when invalid city entered
The city input is an autocomplete field in which it's possible to enter
something besides one of the offered options, but doing so isn't valid and
has been causing the field to be cleared.  This adds an error message and
changes the text of the form hint when the user exits the field with an
invalid chioce, to make it more explicit that they need to pick from the
offered cities.  Since it's setting a form error which disables saving, this
also changes the behavior to leave the invalid entry in the field, since
clearing it produces a weird user experience.

Also fixes the error display for the Organization field, which was mistakenly
referring to the City field's form control object.

Resolves #766.