Prevent red flash when selecting city
The city field has a function that's run on blur to make sure that the
user has chosen from the list of available cities. Unfortunately selecting
a city with a mouse means clicking outside the input, which causes the
onBlur callback to fire before the onSelect callback, while the form value
is still potentially only a portion of the city name. This adds a timeout
to the onBlur callback so that in cases where the user is selecting an item,
the onSelect callback runs first.