Harmonize behavior of multiselect save events
There are a number of actions that cause the multiselect field to save:
- Select from list (with mouse, enter key, or tab key), which should save
  the selected value from the list.
- Enter key with no suggestions showing, which should save exactly what's
  entered in the field.
- Clicking the plus botton, which should save exactly what's in the field,
  even if there are suggestions showing.

There's a bad interaction between the first two, since hitting "enter" when
there are suggestions showing triggers both the "keyup.enter" event and the
typeahead selection event.  This breaks out some of the behavior of the
`add` method to enable different treatment for the different entry methods.
Most notably, it lets us put a timeout on the enter key event so that if the
select event is also firing, it can go first and clear the input to avoid
saving the raw input value in addition to the suggested item.

Connects #721.