Keep unsaved multi-select values
The freeform multiselect field requires that a value be selected from the
list or saved with the enter key or plus button to be actually saved. This
makes a certain amount of sense, but makes it different from most other
fields where typing an answer is sufficient, and it makes it easy to
accidentally lose entries.  This changes it to track all changes and add
values as soon as they're typed.  To preserve the interaction, the 'unsaved'
value is hidden from the list of entered values while the user is on the
page, but it's there in the summary and won't be lost when they leave the
wizard.

Leaving the step and coming back without leaving the wizard has no visible
effect, but leaving the wizard altogether and coming back results in the
unsaved value being converted to a normal saved value and being shown in
the list below the input.

Resolves #771.