Style top concerns with dynamic data.  Resolves #687
The `.concern` cards appear in a number of contexts, some of which
include a title and data, some of which only include a title.
They also appear in two formats: a square format with
vertically centered content, and a compact rectangular format.

I refactored the html a bit and overhauled the scss to achieve a
consistent styles across contexts and formats. I also tweaked the
responsiveness to scale from 1 to 4 columns of cards.