Remove plan icons. Add "Recommended" to 10h consulting plan card.
"Recommended" only appears on the public pricing page, and on the
"manage subscriptions" page during the free trial. It does not appear on
the "manage subscriptions" page if the user already has a paid plan. At
that point, if we're going to recommend a plan, the recommendation should
be based on that user's actual usage and needs.