Use default units for Concern calculations
Logic existed to allow organizations to choose between using Imperial
and Metric units for Concerns, but this was not connected to the UX and
was not fully supported or tested behavior. In the move to storing
Concern values in the database, maintaining a level of half-support for
this feature would require additional infrastructure... it would need to
be possible to determine ahead of time what units should be available
for an indicator, which units correspond to the organization's
preference, and if values for those units exist in the database.

Because of the limited time available before project launch, and because
the ability for organizations to choose between Metric and Imperial is
not currently officially supported, this removes some of the existing
logic to enable cached values to be introduced without requiring that
additional infrastructure.