Disable action wizard fields until action is named
Much like the risk wizard, the form fields of the action wizard should
be disabled until the first step is complete. Users can skip to any step
in the process, but they are prevented from entering any data until they
go back to the first step and give the action a name. The implementation
here is completely based on the implementation of the same behavior in the
risk wizard.

Refs #950