Swap Temperate cache for CCC cache for services
Replaces the Temperate cache with the cache from the CCC library. This
cache service is designed to cache both the in-flight requests and
responses, which reduces the number of API calls made from the
application.

In addition to the services that made use of the Temperate cache, the
CCC cache was also added to the weather event service. This is because
multiple requests were made from the dashboard to this service.

The existing Temperate cache service is left in place because it is
still being utilized by the free trial banner on the dashboard.

Refs #972