Fix indicator charts
In #1025, we switched from using the city object returned from the city
endpoint, to the city object stored with the organization. The city data
stored with an organization does not include the top-level id attribute
or the dataset attribute, both of which are required for the indicator
page.

The changes to the location model not included, this is a temporary fix
for launch. At a later date, we should reintegrate the city service
instead of manually specifying and hardcoding the id and dataset values.
We do not have 'NEX-GDDP' for some cities, so it is not included in the
list of datasets.

Refs #1058