Look for location in Missy's Cities file before geocoding
Since the cities in Missy's data don't change much and we want to be able
to re-do the ingest when we change things, this makes the location/org
creation step look for saved lon/lat values in the cities.csv file.
If no value is found, it falls back to geocoding as before.