Use API city service data for chart city parameter
Follow-up to #1063

The city data returned from the API contains the "dataset" field, which
is needed to fully populate the indicator charts.

We also make the switch to importing the City model from the CCC library
as "ApiCity" for better disambiguation between the Temperate model
Location. These two models were previously conflated.

Refs #1064