Switch Missy Cities geocoder to authenticated Esri
Since we're storing the coordinates in the database, we should be using
a geocoding service that allows that.  Esri does if you supply credentials
and set the `forStorage=true` parameter.  geopy supports Esri but not the
forStorage param, so this:
- switches the geocoder to Esri
- switches the geocoder package from geopy to OMGeo
- adds command-line options for supplyng the the Esri client ID and secret

Since the normal use case will load coordinates from the file and not need
to geocode, the options are optional.  Failing to geocode a city was not a
fatal error before and it still isn't--it will print a warning and move on.