Add state and filter on no API ID for locations in Missy ingest
We want the locations created for the Missy's Dataset ingest to be
independent of the API ones.  We certainly don't want loading the dataset
to fail if the database happens to have imported an API version of one of
the cities in the dataset.  So this adds a filter to the `PlanItLocation`
`update_or_create` call to get only ones with no API ID.

It also converts the state name to the corresponding abbreviation and
saves that on the locations as well, which besides just making sense also
means that the suggested actions will show up with "Locality, ST" rather
than just "Locality".