Refactor UserSerializer to use PrimaryKeyRelatedField
We cannot use the SlugRelatedField because Organization.name
is not sufficient to determine org uniqueness and the docs
explicitly state:

"When using SlugRelatedField as a read-write field, you will
normally want to ensure that the slug field corresponds to a
model field with unique=True."

We weren't doing that. So this swaps to PrimaryKeyRelatedField
and since we now need Org id and name in the client app, we
serialize the entire Organization object in UserOrgSerializer
and use it anywhere we GET a single user object. We continue
to use the default UserSerializer for the list and write
endpoints.

Updates client app to support the API changes as well.