Update front-end to use Location.datasets property
This reverts some (but not all) of the front-end changes made in f8eefbf28.

Contrary to the comment in 'indicator-chart.component.ts', the indicator
charts do NOT gracefully handle the dataset being unavailable, as the error
message asks the user to refresh, which wll never work.

Since the datasets property is now exposed on Location, we can properly
initialize the datasets toggle and avoid this issue.