Lock user DB row when creating organizations
This method as originally written was prone to a classic data race
occurring with multiple workers first checking to see if there was already
an organization and then proceeding to create one only if there was not.

Prior to commit 733b0b0 this possible data race was easiest to encounter
by double clicking on the submit button of the organization creation
wizard.

By locking the user row with select_for_update before checking, we avoid
this data race.