Migrate climate-change-components to temperate project
This allows the temperate Angular app to successfully compile.

PlanItApiHttp was removed in favor of plain old HttpClient.
PlanItApiHttp was causing cryptic errors when requests were
made, such as:
'XMLHttpRequest failed to open with scheme: HTTP://LOCALHOST:8081/API/USER/'
Why caps? And why is that a scheme?

Now we just get normal 401 authorized failures because we need
to implement an HTTP Interceptor to add the token to all API requests.