Indicators endpoint compiles using http4s
But fails due to Out of Memory errors
when a query is made. Then it just hangs
forever attempting to repeat queries. It
never prints any log output, but my
network monitor indicates that the crash
occurs long (>10s) after no more data is
being pulled from S3.

startYear, end Year and area are all stubbed
out for now.

commonsIO dependency needs to exist in the project
but not imported anywhere. Odd.

We'll want to revisit the data structures for
Scenario/Variable/Indicator/ClimateModel later as
part of the tasks to add those endpoints.

Adds developer notes to project READMEs