Add vendored version of difflib.js
This library was failing in AoT mode, due to attempting to assign to the
'.name' property of an anonymous function.

It is unclear to me why this was not an issue before, but is now.

Regardless, since the library was last updated 7 years ago and has several
similarly old unmerged pull requests, it seemed like the best solution was
to copy it into our project and fix it there rather than attempting to
fork it and submit a fix.
to fork it and submit a pull