Update pull request template checklist to include testing the AoT build
It hasn't been a regular part of my process to test the AoT build of the
app served through nginx, though we do have the ability to do so.

In the past this was somewhat annoying, as the Google API is locked down
by referrer, and testing the AoT build would also switch to the production
Google API key, which is not (and should not) be set up for local testing.

Switching away from the Google Maps API makes local testing of the AoT
build easier. We've been bitten enough times by things working in
development but not on staging, that I think it is worth the time
spent to add this to our checklist.