Sending emails was broken on production
Investigation revealed this to be caused by incompatible versions between
transitive dependencies, which were updated to the latest allowed by our
requirements.txt during the last deploy.

This caused us to pull a version of botocore that was not compatible with
our version of urllib3.

Using pipdeptree I determined that we needed to update requests, as our
older version of it was pinned to require an older version of urllib3

I also updated django-amazon-ses at the same time for good measure, to the
latest version still supported by the version of Django/Python we're using

Closes #1424